// frontend/src/api/models/PageResponseProjectResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProjectResponse } from './ProjectResponse';
import {
    ProjectResponseFromJSON,
    ProjectResponseFromJSONTyped,
    ProjectResponseToJSON,
} from './ProjectResponse';

/**
 * 
 * @export
 * @interface PageResponseProjectResponse
 */
export interface PageResponseProjectResponse {
    /**
     * 
     * @type {Array<ProjectResponse>}
     * @memberof PageResponseProjectResponse
     */
    data?: Array<ProjectResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageResponseProjectResponse
     */
    pageNum?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseProjectResponse
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseProjectResponse
     */
    count?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseProjectResponse
     */
    totalPages?: number;
}

/**
 * Check if a given object implements the PageResponseProjectResponse interface.
 */
export function instanceOfPageResponseProjectResponse(value: object): value is PageResponseProjectResponse {
    return true;
}

export function PageResponseProjectResponseFromJSON(json: any): PageResponseProjectResponse {
    return PageResponseProjectResponseFromJSONTyped(json, false);
}

export function PageResponseProjectResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PageResponseProjectResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(ProjectResponseFromJSON)),
        'pageNum': json['pageNum'] == null ? undefined : json['pageNum'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'count': json['count'] == null ? undefined : json['count'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
    };
}

export function PageResponseProjectResponseToJSON(value?: PageResponseProjectResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(ProjectResponseToJSON)),
        'pageNum': value['pageNum'],
        'pageSize': value['pageSize'],
        'count': value['count'],
        'totalPages': value['totalPages'],
    };
}